+++ test/ann/LayerTest.cpp
@@ -23,7 +23,7 @@ LayerTest::LayerTest(QObject *parent) :
 void LayerTest::testLayerCreation()
 {
     Layer layer;
-    QCOMPARE(layer.size(), 0);
+    QCOMPARE(layer.size(), 0ul);
 }
 
 
@@ -33,7 +33,7 @@ void LayerTest::testNeuronAddition()
     layer << new Neuron(new LinearActivationFunction(), this);
     layer << new Neuron(new LinearActivationFunction(), this);
 
-    QCOMPARE(layer.size(), 2);
+    QCOMPARE(layer.size(), 2ul);
     QVERIFY(layer.neuronAt(0)->parent() == &layer);
     QVERIFY(layer.neuronAt(1)->parent() == &layer);
 }
@@ -51,7 +51,7 @@ void LayerTest::testNeuronIterator()
         neurons << neuron;
     });
 
-    QCOMPARE(neurons.size(), layer.size());
+    QCOMPARE(static_cast<size_t>(neurons.size()), layer.size());
 }
